Flexible format

+++ wiegand.h
@@ -1,2 +1,2 @@
 void wiegand_init();
-uint32_t wiegand_read();
+uint64_t wiegand_read();
